+++ aws_xray_sdk/core/models/trace_header.py
@@ -5,6 +5,7 @@ log = logging.getLogger(__name__)
 ROOT = 'Root'
 PARENT = 'Parent'
 SAMPLE = 'Sampled'
+SELF = 'Self'
 
 HEADER_DELIMITER = ";"
 
@@ -17,18 +18,25 @@ class TraceHeader(object):
     by the X-Ray SDK and included in the response. Learn more about
     `Tracing Header <http://docs.aws.amazon.com/xray/latest/devguide/xray-concepts.html#xray-concepts-tracingheader>`_.
     """
-    def __init__(self, root=None, parent=None, sampled=None):
+    def __init__(self, root=None, parent=None, sampled=None, data=None):
         """
         :param str root: trace id
         :param str parent: parent id
         :param int sampled: 0 means not sampled, 1 means sampled
+        :param dict data: arbitrary data fields
         """
         self._root = root
         self._parent = parent
         self._sampled = None
+        self._data = data
 
         if sampled is not None:
-            self._sampled = int(sampled)
+            if sampled == '?':
+                self._sampled = sampled
+            if sampled is True or sampled == '1' or sampled == 1:
+                self._sampled = 1
+            if sampled is False or sampled == '0' or sampled == 0:
+                self._sampled = 0
 
     @classmethod
     def from_header_str(cls, header):
@@ -42,15 +50,22 @@ class TraceHeader(object):
         try:
             params = header.strip().split(HEADER_DELIMITER)
             header_dict = {}
+            data = {}
 
             for param in params:
                 entry = param.split('=')
-                header_dict[entry[0]] = entry[1]
+                key = entry[0]
+                if key in (ROOT, PARENT, SAMPLE):
+                    header_dict[key] = entry[1]
+                # Ignore any "Self=" trace ids injected from ALB.
+                elif key != SELF:
+                    data[key] = entry[1]
 
             return cls(
                 root=header_dict.get(ROOT, None),
                 parent=header_dict.get(PARENT, None),
                 sampled=header_dict.get(SAMPLE, None),
+                data=data,
             )
 
         except Exception:
@@ -62,15 +77,18 @@ class TraceHeader(object):
         Convert to a tracing header string that can be injected to
         outgoing http request headers.
         """
-        h_str = ''
+        h_parts = []
         if self.root:
-            h_str = ROOT + '=' + self.root
+            h_parts.append(ROOT + '=' + self.root)
         if self.parent:
-            h_str = h_str + HEADER_DELIMITER + PARENT + '=' + self.parent
+            h_parts.append(PARENT + '=' + self.parent)
         if self.sampled is not None:
-            h_str = h_str + HEADER_DELIMITER + SAMPLE + '=' + str(self.sampled)
+            h_parts.append(SAMPLE + '=' + str(self.sampled))
+        if self.data:
+            for key in self.data:
+                h_parts.append(key + '=' + self.data[key])
 
-        return h_str
+        return HEADER_DELIMITER.join(h_parts)
 
     @property
     def root(self):
@@ -90,6 +108,13 @@ class TraceHeader(object):
     def sampled(self):
         """
         Return the sampling decision in the header.
-        It's either 0 or 1.
+        It's 0 or 1 or '?'.
         """
         return self._sampled
+
+    @property
+    def data(self):
+        """
+        Return the arbitrary fields in the trace header.
+        """
+        return self._data