+++ aws_xray_sdk/ext/httplib/patch.py
@@ -20,6 +20,8 @@ else:
 
 _XRAY_PROP = '_xray_prop'
 _XRay_Data = namedtuple('xray_data', ['method', 'host', 'url'])
+# A flag indicates whether this module is X-Ray patched or not
+PATCH_FLAG = '__xray_patched'
 
 
 def http_response_processor(wrapped, instance, args, kwargs, return_value,
@@ -42,7 +44,8 @@ def http_response_processor(wrapped, instance, args, kwargs, return_value,
 
 def _xray_traced_http_getresponse(wrapped, instance, args, kwargs):
     if not PY2 and kwargs.get('buffering', False):
-        return wrapped(*args, **kwargs)  # ignore py2 calls that fail as 'buffering` only exists in py2.
+        # ignore py2 calls that fail as 'buffering` only exists in py2.
+        return wrapped(*args, **kwargs)
 
     xray_data = getattr(instance, _XRAY_PROP)
 
@@ -70,7 +73,8 @@ def _send_request(wrapped, instance, args, kwargs):
     def decompose_args(method, url, body, headers, encode_chunked=False):
         inject_trace_header(headers, xray_recorder.current_subsegment())
 
-        # we have to check against sock because urllib3's HTTPSConnection inherit's from http.client.HTTPConnection
+        # we have to check against sock because urllib3's HTTPSConnection
+        # inherits from `http.client.HTTPConnection`.
         scheme = 'https' if isinstance(instance.sock, ssl.SSLSocket) else 'http'
         xray_url = '{}://{}{}'.format(scheme, instance.host, url)
         xray_data = _XRay_Data(method, instance.host, xray_url)
@@ -101,7 +105,9 @@ def http_read_processor(wrapped, instance, args, kwargs, return_value,
 
 
 def _xray_traced_http_client_read(wrapped, instance, args, kwargs):
-    xray_data = getattr(instance, _XRAY_PROP)
+    xray_data = getattr(instance, _XRAY_PROP, None)
+    if not xray_data:
+        return wrapped(*args, **kwargs)
 
     return xray_recorder.record_subsegment(
         wrapped, instance, args, kwargs,
@@ -112,12 +118,13 @@ def _xray_traced_http_client_read(wrapped, instance, args, kwargs):
 
 
 def patch():
-    """ patch the built-in urllib/httplib/httplib.client methods for tracing"""
-
-    # we set an attribute to avoid double unwrapping
-    if getattr(httplib, '__xray_patch', False):
+    """
+    patch the built-in `urllib/httplib/httplib.client` methods for tracing.
+    """
+    if getattr(httplib, PATCH_FLAG, False):
         return
-    setattr(httplib, '__xray_patch', True)
+    # we set an attribute to avoid multiple wrapping
+    setattr(httplib, PATCH_FLAG, True)
 
     wrapt.wrap_function_wrapper(
         httplib_client_module,
@@ -139,13 +146,12 @@ def patch():
 
 
 def unpatch():
-    """ unpatch any previously patched modules """
-    if not getattr(httplib, '__xray_patch', False):
-        return
-    setattr(httplib, '__xray_patch', False)
-
-    # send_request encapsulates putrequest, putheader[s], and endheaders
-    # NOTE that requests
+    """
+    Unpatch any previously patched modules.
+    This operation is idempotent.
+    """
+    setattr(httplib, PATCH_FLAG, False)
+    # _send_request encapsulates putrequest, putheader[s], and endheaders
     unwrap(httplib.HTTPConnection, '_send_request')
     unwrap(httplib.HTTPConnection, 'getresponse')
-    unwrap(httplib.HTTPConnection, 'read')
+    unwrap(httplib.HTTPResponse, 'read')