+++ tests/ext/botocore/test_botocore.py
@@ -50,6 +50,30 @@ def test_ddb_table_name():
     assert aws_meta['operation'] == 'DescribeTable'
 
 
+def test_s3_bucket_name_capture():
+    s3 = session.create_client('s3', region_name='us-west-2')
+    response = {
+        'ResponseMetadata': {
+            'RequestId': REQUEST_ID,
+            'HTTPStatusCode': 200,
+        }
+    }
+
+    bucket_name = 'mybucket'
+
+    with Stubber(s3) as stubber:
+        stubber.add_response('list_objects_v2', response, {'Bucket': bucket_name})
+        s3.list_objects_v2(Bucket=bucket_name)
+
+    subsegment = xray_recorder.current_segment().subsegments[0]
+    aws_meta = subsegment.aws
+
+    assert aws_meta['bucket_name'] == bucket_name
+    assert aws_meta['request_id'] == REQUEST_ID
+    assert aws_meta['region'] == 'us-west-2'
+    assert aws_meta['operation'] == 'ListObjectsV2'
+
+
 def test_list_parameter_counting():
     """
     Test special parameters that have shape of list are recorded
@@ -103,3 +127,26 @@ def test_map_parameter_grouping():
 
     aws_meta = subsegment.aws
     assert sorted(aws_meta['table_names']) == ['table1', 'table2']
+
+def test_pass_through_on_context_missing():
+    """
+    The built-in patcher or subsegment capture logic should not throw
+    any error when a `None` subsegment created from `LOG_ERROR` missing context.
+    """
+    xray_recorder.configure(context_missing='LOG_ERROR')
+    xray_recorder.clear_trace_entities()
+
+    ddb = session.create_client('dynamodb', region_name='us-west-2')
+    response = {
+        'ResponseMetadata': {
+            'RequestId': REQUEST_ID,
+            'HTTPStatusCode': 200,
+        }
+    }
+
+    with Stubber(ddb) as stubber:
+        stubber.add_response('describe_table', response, {'TableName': 'mytable'})
+        result = ddb.describe_table(TableName='mytable')
+    assert result is not None
+
+    xray_recorder.configure(context_missing='RUNTIME_ERROR')