+++ tests/ext/sqlalchemy/test_query.py
@@ -42,7 +42,7 @@ def test_all(capsys, session):
     session.query(User).all()
     subsegment = find_subsegment_by_annotation(xray_recorder.current_segment(), 'sqlalchemy', 'sqlalchemy.orm.query.all')
     assert subsegment['annotations']['sqlalchemy'] == 'sqlalchemy.orm.query.all'
-    # assert subsegment['sql']['sanitized_query']
+    assert subsegment['sql']['sanitized_query']
     assert subsegment['sql']['url']
 
 
@@ -57,13 +57,14 @@ def test_add(capsys, session):
     assert subsegment['sql']['url']
 
 
-def test_filter(capsys, session):
-    """ Test calling all() on get all records.
+def test_filter_first(capsys, session):
+    """ Test calling filter().first() on get first filtered records.
     Verify we run the query and return the SQL as metdata"""
     # with capsys.disabled():
-    session.query(User).filter(User.password=="mypassword!")
-    subsegment = find_subsegment_by_annotation(xray_recorder.current_segment(), 'sqlalchemy', 'sqlalchemy.orm.query.filter')
-    assert subsegment['annotations']['sqlalchemy'] == 'sqlalchemy.orm.query.filter'
-    # assert subsegment['sql']['sanitized_query']
-    # assert "mypassword!" not in subsegment['sql']['sanitized_query']
+    session.query(User).filter(User.password=="mypassword!").first()
+    subsegment = find_subsegment_by_annotation(xray_recorder.current_segment(), 'sqlalchemy', 'sqlalchemy.orm.query.first')
+    assert subsegment['annotations']['sqlalchemy'] == 'sqlalchemy.orm.query.first'
+    assert subsegment['sql']['sanitized_query']
+    assert "mypassword!" not in subsegment['sql']['sanitized_query']
+    assert "users.password = ?" in  subsegment['sql']['sanitized_query']
     assert subsegment['sql']['url']